--- src/main/java/com/slavisa/BudgetApp/Main.java
@@ -1,13 +1,0 @@
-package com.slavisa.BudgetApp;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Main {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Main.class, args);
-	}
-
-}
