--- src/main/java/com/slavisa/BudgetApp/auth/AuthenticationController.java
@@ -1,21 +1,0 @@
-package com.slavisa.BudgetApp.auth;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/login")
-public class AuthenticationController {
-
-    @Autowired
-    private AuthenticationService service;
-
-    @PostMapping()
-    public ResponseEntity<AuthenticationResponse> login(@RequestBody AuthenticationRequest request){
-        return ResponseEntity.ok(service.authenticate(request));
-    }
-}
