--- src/main/java/com/slavisa/BudgetApp/auth/AuthenticationRequest.java
@@ -1,16 +1,0 @@
-package com.slavisa.BudgetApp.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthenticationRequest {
-
-    private String username;
-    private String password;
-}
