--- src/main/java/com/slavisa/BudgetApp/auth/AuthenticationResponse.java
@@ -1,15 +1,0 @@
-package com.slavisa.BudgetApp.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationResponse {
-
-    private String token;
-}
