--- src/main/java/com/slavisa/BudgetApp/auth/AuthenticationService.java
@@ -1,36 +1,0 @@
-package com.slavisa.BudgetApp.auth;
-
-import com.slavisa.BudgetApp.config.JwtService;
-import com.slavisa.BudgetApp.repository.AccountRepo;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.stereotype.Service;
-
-@Service
-public class AuthenticationService {
-
-    @Autowired
-    private AccountRepo accountRepo;
-
-    @Autowired
-    private AuthenticationManager authenticationManager;
-
-    @Autowired
-    private JwtService jwtService;
-
-
-    public AuthenticationResponse authenticate(AuthenticationRequest request) {
-        authenticationManager.authenticate(
-          new UsernamePasswordAuthenticationToken(
-                  request.getUsername(),
-                  request.getPassword()
-          )
-        );
-        var user = accountRepo.findByUsername(request.getUsername()).orElseThrow();
-        var jwtToken = jwtService.generateToken(user);
-        return AuthenticationResponse.builder()
-                .token(jwtToken)
-                .build();
-    }
-}
