--- src/main/java/com/slavisa/BudgetApp/config/JwtService.java
@@ -1,77 +1,0 @@
-package com.slavisa.BudgetApp.config;
-
-import com.slavisa.BudgetApp.model.Account;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Service;
-
-import java.security.Key;
-import java.util.Date;
-import java.util.function.Function;
-
-@Service
-public class JwtService {
-
-    @Value("${jwt.secret}")
-    private String SECRET;
-
-
-    public String extractUsername(String token) {
-        return extractClaim(token, Claims::getSubject);
-    }
-
-    public <T> T extractClaim(String token, Function<Claims, T> claimsResolver){
-        final Claims claims= extractAllClaims(token);
-        return claimsResolver.apply(claims);
-    }
-
-    public String generateToken( UserDetails userDetails){
-        Integer accountID = ((Account) userDetails).getAccountID();
-        String role = userDetails.getAuthorities()
-                .iterator()
-                .next()
-                .getAuthority();
-
-        return Jwts
-                .builder()
-                .subject(userDetails.getUsername())
-                .claim("role",role)
-                .claim("accountID",accountID)
-                .issuedAt(new Date(System.currentTimeMillis()))
-                .expiration(new Date(System.currentTimeMillis() + 1000*60*30))
-                .signWith(getSignInKey(), SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public boolean isTokenValid(String token, UserDetails userDetails){
-        final String username = extractUsername(token);
-        return (username.equals(userDetails.getUsername())) && !isTokenExpired(token);
-    }
-
-    private boolean isTokenExpired(String token) {
-        return extractExpiration(token).before(new Date());
-    }
-
-    private Date extractExpiration(String token) {
-        return extractClaim(token,Claims::getExpiration);
-    }
-
-    private Claims extractAllClaims(String token){
-        return Jwts
-                .parser()
-                .setSigningKey(getSignInKey())
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-    private Key getSignInKey() {
-        byte[] keyBytes= Decoders.BASE64.decode(SECRET);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-}
