--- src/main/java/com/slavisa/BudgetApp/model/Role.java
@@ -1,28 +1,0 @@
-package com.slavisa.BudgetApp.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Entity
-@Data
-@NoArgsConstructor
-public class Role {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer roleID;
-
-    @NotBlank
-    @Size(max = 10, message = "Maximum is 10 characters.")
-    private String roleName;
-
-    @OneToMany(mappedBy = "role")
-    @JsonIgnore
-    private List<Account> users;
-}
