--- src/main/java/com/slavisa/BudgetApp/model/Transaction.java
@@ -1,45 +1,0 @@
-package com.slavisa.BudgetApp.model;
-
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-@Entity
-@Data
-@NoArgsConstructor
-@Table(name = "Transactions")
-public class Transaction {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer transactionID;
-
-    @ManyToOne
-    @JoinColumn(name = "AccountID")
-    private Account account;
-
-    @ManyToOne
-    @JoinColumn(name = "CategoryID")
-    private Category category;
-
-    @NotBlank
-    @Size(max = 255, message = "Maximum is 255 characters.")
-    private String description;
-
-    @Column(nullable = false, precision = 10, scale = 2)
-    private BigDecimal amount;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private TransactionType type;
-
-    @Column(nullable = false)
-    private LocalDate date;
-}
