--- src/main/java/com/slavisa/BudgetApp/model/TransactionType.java
@@ -1,6 +1,0 @@
-package com.slavisa.BudgetApp.model;
-
-public enum TransactionType {
-    INCOME,
-    EXPENSE
-}
