--- src/main/java/com/slavisa/BudgetApp/repository/AccountRepo.java
@@ -1,13 +1,0 @@
-package com.slavisa.BudgetApp.repository;
-
-import com.slavisa.BudgetApp.model.Account;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface AccountRepo extends JpaRepository<Account,Integer> {
-
-    Optional<Account> findByUsername(String name);
-}
