--- src/main/java/com/slavisa/BudgetApp/repository/CategoryRepo.java
@@ -1,9 +1,0 @@
-package com.slavisa.BudgetApp.repository;
-
-import com.slavisa.BudgetApp.model.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepo extends JpaRepository<Category,Integer> {
-}
