--- src/main/java/com/slavisa/BudgetApp/repository/RoleRepo.java
@@ -1,9 +1,0 @@
-package com.slavisa.BudgetApp.repository;
-
-import com.slavisa.BudgetApp.model.Role;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RoleRepo extends JpaRepository<Role,Integer> {
-}
