--- src/main/java/com/slavisa/BudgetApp/service/RoleService.java
@@ -1,39 +1,0 @@
-package com.slavisa.BudgetApp.service;
-
-import com.slavisa.BudgetApp.model.Role;
-import com.slavisa.BudgetApp.repository.RoleRepo;
-import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-public class RoleService {
-
-    @Autowired
-    private RoleRepo roleRepo;
-
-    public List<Role> getAllRoles(){
-        return roleRepo.findAll();
-    }
-
-    public Optional<Role> getRoleById(Integer roleID){
-        return roleRepo.findById(roleID);
-    }
-
-    public void deleteRoleByID(Integer roleID){
-        roleRepo.deleteById(roleID);
-    }
-
-    public Role updateRole(Integer roleID, Role role){
-        Role role1 = roleRepo.findById(roleID).orElseThrow(() -> new RuntimeException("Role doesn't exist."));
-        role1.setRoleName(role.getRoleName());
-        return roleRepo.save(role1);
-    }
-
-    public Role addRole(@Valid Role role) {
-        return roleRepo.save(role);
-    }
-}
