+++ src/main/java/com/slavisa/budgetapp/auth/AuthenticationRequest.java
@@ -1,0 +1,16 @@
+package com.slavisa.budgetapp.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthenticationRequest {
+
+    private String username;
+    private String password;
+}
