+++ src/main/java/com/slavisa/budgetapp/auth/AuthenticationResponse.java
@@ -1,0 +1,15 @@
+package com.slavisa.budgetapp.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationResponse {
+
+    private String token;
+}
