+++ src/main/java/com/slavisa/budgetapp/model/Transaction.java
@@ -1,0 +1,44 @@
+package com.slavisa.budgetapp.model;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+import java.math.BigDecimal;
+import java.time.LocalDate;
+
+@Entity
+@Data
+@NoArgsConstructor
+@Table(name = "Transactions")
+public class Transaction implements Serializable {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer transactionID;
+
+    @ManyToOne
+    @JoinColumn(name = "AccountID")
+    private Account account;
+
+    @ManyToOne
+    @JoinColumn(name = "CategoryID")
+    private Category category;
+
+    @NotBlank
+    @Size(max = 255, message = "Maximum is 255 characters.")
+    private String description;
+
+    @Column(nullable = false, precision = 10, scale = 2)
+    private BigDecimal amount;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private TransactionType type;
+
+    @Column(nullable = false)
+    private LocalDate date;
+}
