+++ src/main/java/com/slavisa/budgetapp/model/TransactionType.java
@@ -1,0 +1,6 @@
+package com.slavisa.budgetapp.model;
+
+public enum TransactionType {
+    INCOME,
+    EXPENSE
+}
