+++ src/main/java/com/slavisa/budgetapp/repository/CategoryRepo.java
@@ -1,0 +1,9 @@
+package com.slavisa.budgetapp.repository;
+
+import com.slavisa.budgetapp.model.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryRepo extends JpaRepository<Category,Integer> {
+}
