+++ src/main/java/com/slavisa/budgetapp/repository/RoleRepo.java
@@ -1,0 +1,9 @@
+package com.slavisa.budgetapp.repository;
+
+import com.slavisa.budgetapp.model.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RoleRepo extends JpaRepository<Role,Integer> {
+}
