--- src/test/java/com/slavisa/BudgetApp/BudgetAppApplicationTests.java
@@ -1,13 +1,0 @@
-package com.slavisa.BudgetApp;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BudgetAppApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
