+++ src/test/java/com/slavisa/budgetapp/BudgetAppApplicationTests.java
@@ -1,0 +1,14 @@
+package com.slavisa.budgetapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BudgetAppApplicationTests {
+
+	@Test
+	void contextLoads() {
+        //
+	}
+
+}
